requirements.txt
# Framework
Flask==2.3.2
Werkzeug==2.3.6
flask-cors==4.0.0

# Database
SQLAlchemy==2.0.19

# Cache & Queue
redis==4.6.0
pika==1.3.1

# Image Processing
Pillow==11.0.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Logging & Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Development
black==23.7.0
flake8==6.0.0
